ManageabilityPkg: Check PLDM completion code
Every PLDM response contains 1 byte Completion code which is needed
to be checked against SUCCESS value which means that the PLDM command
was accepted and completed normally.
If it is not equal to that value it is necessary to return error status
so the caller would know that Response buffer is invalid.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>